[expo-modules][android] Extract `ModuleDestroyedException` into `expo-modules-core`
