--- packages/expo-barcode-scanner/android/src/main/java/expo/modules/barcodescanner/ScopeCancelationException.kt
@@ -1,5 +1,0 @@
-package expo.modules.barcodescanner
-
-import kotlinx.coroutines.CancellationException
-
-class ScopeCancellationException : CancellationException("View destroyed, scope canceled")
