--- packages/expo-image-picker/android/src/main/java/expo/modules/imagepicker/ModuleDestroyedException.kt
@@ -1,5 +1,0 @@
-package expo.modules.imagepicker
-
-import kotlinx.coroutines.CancellationException
-
-class ModuleDestroyedException : CancellationException(ImagePickerConstants.PROMISES_CANCELED)
