+++ packages/expo-image/android/src/main/java/expo/modules/image/ExpoImageModule.kt
@@ -1,5 +1,6 @@
 package expo.modules.image
 
+import android.util.Log
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.model.GlideUrl
 import com.bumptech.glide.request.FutureTarget
@@ -9,9 +10,12 @@ import com.facebook.react.bridge.ReactContextBaseJavaModule
 import com.facebook.react.bridge.ReactMethod
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.cancel
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runInterruptible
 import java.lang.Exception
+import java.lang.IllegalStateException
+import java.util.concurrent.CancellationException
 
 /**
  * We need to convert blocking java.util.concurrent.Future result
@@ -41,5 +45,16 @@ class ExpoImageModule(val context: ReactApplicationContext) : ReactContextBaseJa
         promise.reject("ERR_IMAGE_PREFETCH_FAILURE", "Failed to prefetch the image: ${e.message}", e)
       }
     }
+  }
+
+  override fun onCatalystInstanceDestroy() {
+    try {
+      // TODO: Use [expo.modules.core.errors.ModuleDestroyedException] when migrated to Expo Module
+      moduleCoroutineScope.cancel(CancellationException("ExpoImage module is destroyed. Cancelling all jobs."))
+    } catch (e: IllegalStateException) {
+      Log.w("ExpoImageModule", "No coroutines to cancel")
+    }
+
+    super.onCatalystInstanceDestroy()
   }
 }
